refactor(chat): centralize Gimli-AI state management with global atom
- Move Gimli-AI choice logic from local component state to gimliChoiceAtom
- Remove useGimliChoice custom hook from thread component
- Replace useRef implementation with global atom in FakeAssistantMessage
- Update AssistantMessage to use centralized state management
- Improve state consistency across component re-renders
